Formatting and adding a new way of handling columns
My tables had extra columns that did not belong to my ORM models, so
when Paginate was getting these columns it would throw an error.
Now you can pass an array of columns to the execute function, like this
:
$columns = array(
            array(
                'key' => 'not_from_my_model',
                'name'=>'My friendly name',
                'paginate' => false, //don't use for paginate
                'filter'=>false, //don't use for filtering
            ),
            array(
                'key' => 'my_model_name.my_columns_name',
                'name'=> 'My other friendly name',
            ),
        );

Thus when calling datatables for ordering/filtering/searching we can
compare the different columns use for pagination and in the view and use
only the relvant one.
Work still in progress...